feat: extract logs to diagonstics screen
