+++ src/pages/SettingsViewV2/content/ReportAProblemContent/index.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react'
+import React, { useCallback, useState } from 'react'
 
 import {
   sendGoogleFormFeedback,
@@ -8,10 +8,9 @@ import {
   Button,
   Form,
   TextArea,
-  PageHeader,
-  ToggleSwitch
+  PageHeader
 } from '@tetherto/pearpass-lib-ui-kit'
-import { FolderOpen, Send } from '@tetherto/pearpass-lib-ui-kit/icons'
+import { Send } from '@tetherto/pearpass-lib-ui-kit/icons'
 
 import {
   GOOGLE_FORM_KEY,
@@ -33,9 +32,7 @@ const OFFLINE_TIMEOUT_MESSAGE =
 const TEST_IDS = {
   root: 'settings-card-report',
   textarea: 'settings-report-textarea',
-  send: 'settings-report-send-button',
-  openLogs: 'settings-report-open-logs-button',
-  loggingToggle: 'settings-report-logging-toggle'
+  send: 'settings-report-send-button'
 } as const
 
 type ReportAProblemContentProps = {
@@ -51,50 +48,8 @@ export const ReportAProblemContent = ({
 
   const [message, setMessage] = useState('')
   const [isLoading, setIsLoading] = useState(false)
-  const [loggingEnabled, setLoggingEnabled] = useState(false)
-  const [loggingForced, setLoggingForced] = useState(false)
-  const [isTogglingLogging, setIsTogglingLogging] = useState(false)
 
   useGlobalLoading({ isLoading })
-
-  useEffect(() => {
-    const electronAPI = window.electronAPI
-    if (!electronAPI || typeof electronAPI.isLoggingEnabled !== 'function') {
-      return
-    }
-
-    let cancelled = false
-    electronAPI
-      .isLoggingEnabled()
-      .then((state) => {
-        if (cancelled) return
-        setLoggingEnabled(state.enabled)
-        setLoggingForced(state.forced)
-      })
-      .catch((error) =>
-        logger.error('ReportAProblemContent', 'isLoggingEnabled failed:', error)
-      )
-    return () => {
-      cancelled = true
-    }
-  }, [])
-
-  const handleLoggingToggle = useCallback(
-    async (next: boolean) => {
-      if (loggingForced || isTogglingLogging) return
-      setIsTogglingLogging(true)
-      try {
-        const state = await window.electronAPI?.setLogging?.(next)
-        if (state) {
-          setLoggingEnabled(state.enabled)
-          setLoggingForced(state.forced)
-        }
-      } finally {
-        setIsTogglingLogging(false)
-      }
-    },
-    [isTogglingLogging, loggingForced]
-  )
 
   const handleSend = useCallback(async () => {
     if (!message?.trim() || isLoading) {
@@ -208,40 +163,6 @@ export const ReportAProblemContent = ({
           </Button>
         </div>
       </Form>
-      <div style={styles.diagnostics}>
-        <ToggleSwitch
-          data-testid={TEST_IDS.loggingToggle}
-          checked={loggingEnabled}
-          onChange={(checked) => {
-            void handleLoggingToggle(checked)
-          }}
-          disabled={loggingForced || isTogglingLogging}
-          label={t('Enable diagnostic logging')}
-          description={
-            loggingForced
-              ? t(
-                  'Logging is enabled by this build (nightly or --enable-logging launch flag).'
-                )
-              : t(
-                  'Diagnostic logs help us troubleshoot issues. Enable, reproduce the problem, then share the logs with us.'
-                )
-          }
-        />
-        <div style={styles.actions}>
-          <Button
-            data-testid={TEST_IDS.openLogs}
-            variant="primary"
-            size="small"
-            disabled={!loggingEnabled}
-            onClick={() => {
-              void window.electronAPI?.openLogsFolder?.()
-            }}
-            iconBefore={<FolderOpen />}
-          >
-            {t('Open logs folder')}
-          </Button>
-        </div>
-      </div>
     </div>
   )
 }
