+++ src/pages/SettingsViewV2/content/index.ts
@@ -1,6 +1,7 @@
 import { AppPreferencesContent } from './AppPreferencesContent'
 import { AppVersionContent } from './AppVersionContent'
 import { BlindPeersContent } from './BlindPeersContent'
+import { DiagnosticsContent } from './DiagnosticsContent'
 import { ExportItemsContent } from './ExportItemsContent'
 import { ImportItemsContent } from './ImportItemsContent'
 import { LanguageContent } from './LanguageContent'
@@ -12,6 +13,7 @@ export {
   AppPreferencesContent,
   AppVersionContent,
   BlindPeersContent,
+  DiagnosticsContent,
   ExportItemsContent,
   ImportItemsContent,
   YourDevicesContent,
